fixed: manual loading of VobSub subtitles
Allow loading VobSub subtitles manually by specifying either the .idx
or the .sub file. The other file is detected automatically. The
following naming schemes are supported:
- foobar.idx and foobar.sub
- foobar.idx and foobar.rar containing foobar.sub

Note that previously CDVDPlayer::AddSubtitleFile() tried to ignore
VobSub .sub files. This was presumably in order to not autoload VobSub
files twice (.sub and .idx), but this is already handled by the caller
by filtering out the VobSub .sub.

The changes are also applied to AMLPlayer and OMXPlayer.